+++ jax/interpreters/ad.py
@@ -15,7 +15,7 @@
 
 import functools
 import itertools as it
-from typing import Any, Callable, Dict
+from typing import Any, Callable, Dict, Set, List
 
 from . import partial_eval as pe
 from .. import core as core
@@ -198,9 +198,18 @@ def backward_pass(jaxpr: core.Jaxpr, consts, primals_in, cotangents_in):
                                      map(read_primal, eqn.invars), params)
         map(write_primal, eqn.outvars, ans)
 
+  # Find the last use of each cotangent so that they can be removed
+  # as soon as possible.
+  drop_cts: List[Set[Any]] = []
+  seen_vars: Set[Any] = set(jaxpr.invars)
+  for eqn in linear_eqns:
+    read_set = set(eqn.outvars)  # NOTE: eqn is not transposed yet!
+    drop_cts.append(read_set - seen_vars)
+    seen_vars |= read_set
+
   ct_env: Dict[Any, Any] = {}
   map(write_cotangent, jaxpr.outvars, cotangents_in)
-  for eqn in linear_eqns[::-1]:
+  for eqn, to_drop in zip(linear_eqns[::-1], drop_cts[::-1]):
     # FIXME: Some invars correspond to tangents
     invals = map(read_primal, eqn.invars)
     if eqn.primitive.multiple_results:
@@ -216,6 +225,8 @@ def backward_pass(jaxpr: core.Jaxpr, consts, primals_in, cotangents_in):
     cts_out = [zero] * len(eqn.invars) if cts_out is zero else cts_out
     # FIXME: Some invars correspond to primals!
     map(write_cotangent, eqn.invars, cts_out)
+    for var in to_drop:
+      ct_env.pop(var, None)  # NB: Constant cotangents might be missing
 
   cotangents_out = map(read_cotangent, jaxpr.invars)
   return cotangents_out